working app

+++ zad1/app/Http/Controllers/NoteController.php
@@ -5,14 +5,17 @@ namespace App\Http\Controllers;
 use App\Models\Note;
 use Illuminate\Http\Request;
 
-class noteController extends Controller
+class NoteController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
-        $notes = Note::query() -> orderBy('created_at', 'desc') -> paginate (10);
+        $notes = Note::query() 
+       // -> where('user_id', request()-> user()->id)
+        -> orderBy('created_at', 'desc') 
+        -> paginate (10);
        // dd($notes);
         return view('note.index', ['notes' => $notes]);
     }
@@ -35,7 +38,7 @@ class noteController extends Controller
             ]);
 
 
-        $data['user_id'] = 1;
+        $data['user_id'] = $request->user()->id;
 
        $note = Note::create($data);
 
@@ -47,6 +50,10 @@ class noteController extends Controller
      */
     public function show(Note $note)
     {
+        if($note->user_id !== request()->user()->id)
+        {
+            abort(403);
+        }
         return  view('note.show', ['note' => $note]);
     }
 
@@ -55,6 +62,10 @@ class noteController extends Controller
      */
     public function edit(Note $note)
     {
+        if($note->user_id !== request()->user()->id)
+        {
+            abort(403);
+        }
         return  view('note.edit', ['note' => $note]);
     }
 
@@ -63,6 +74,11 @@ class noteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
+        if($note->user_id !== request()->user()->id)
+        {
+            abort(403);
+        }
+    
         $data  = $request->validate([
             'note' => ['required', 'string']
             ]);
@@ -77,6 +93,11 @@ class noteController extends Controller
      */
     public function destroy(Note $note)
     {
+        if($note->user_id !== request()->user()->id)
+        {
+            abort(403);
+        }
+
         $note->delete();
 
        return to_route('note.index')->with('message', 'Note was deleted');
